feat: add modal for ctpat inspection close CAR-11
