+++ next.config.mjs
@@ -7,16 +7,6 @@ await import("./src/env.mjs")
 /** @type {import("next").NextConfig} */
 const config = {
   reactStrictMode: true,
-
-  /**
-   * If you are using `appDir` then you must comment the below `i18n` config out.
-   *
-   * @see https://github.com/vercel/next.js/issues/41980
-   */
-  i18n: {
-    locales: ["en"],
-    defaultLocale: "en"
-  },
   swcMinify: true,
   experimental: {
     serverActions: true
